move methods countOf() and isSatisfiedBy() to EntitySpecificationRepositoryCounterTrait
